requirements.txt
flask
flask_bcrypt
requests
flask_sqlalchemy
pymysql
python-dotenv
Flask-Migrate